feat(frontend): extract ydata sections and App into components
